Refactor `GetDeclarativeDocuments.kt`: split files, extract the model
