+++ gradle-client/src/jvmMain/kotlin/org/gradle/client/ui/connected/ConnectedComponent.kt
@@ -12,6 +12,7 @@ import org.gradle.client.core.gradle.GradleConnectionParameters
 import org.gradle.client.core.gradle.GradleDistribution
 import org.gradle.client.ui.AppDispatchers
 import org.gradle.client.ui.connected.actions.*
+import org.gradle.client.ui.connected.actions.declarativedocuments.GetDeclarativeDocuments
 import org.gradle.tooling.GradleConnector
 import org.gradle.tooling.ProjectConnection
 import org.gradle.tooling.events.OperationType
